+++ ursa_sharing/src/pedersen.rs
@@ -104,7 +104,7 @@ impl Scheme {
             return Err(SharingError::PedersenBlindShareInvalid);
         }
 
-        let x = S::from_usize(share.identifier);
+        let x = S::from_usize(share.identifier as usize);
         let mut i = S::one();
 
         // FUTURE: execute this sum of products
